fix spec object creation in arrays
if you access an index in an array that doesn't exist, jqawk now creates
a spec object at that index instead of a null value. that makes
expressions like `a[0].thing = 2` work, if `a[0]` doesn't exist.

this works by putting the array index in the `Num` field on the specobj.