add json input to the fuzz tests
i totally forgot the json input to the fuzz tests was hardcoded. this
adds a new fuzz test that also fuzzes the json. for this to work, i need
to wrap JSON parsing errors, so the fuzzer knows they're not crashes. it
also means i format them a bit nicer too.